% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-data-utils.R
\name{assay_list_file_path}
\alias{assay_list_file_path}
\title{Returns the filepath for the raw assay data for a given assay(name)}
\usage{
assay_list_file_path(assay_name)
}
\arguments{
\item{assay_name}{One of the (case insensitive) options in
`available_assays()$assay_name`}
}
\value{
file path to the serialized assaylist
}
\description{
Returns the filepath for the raw assay data for a given assay(name)
}
