% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-data-utils.R
\name{assay_lists_load}
\alias{assay_lists_load}
\title{Loads an assay list object}
\usage{
assay_lists_load(assay_name)
}
\arguments{
\item{assay_name}{One of the (case insensitive) options in
`available_assays()$assay_name`}
}
\description{
Loads an assay list object
}
