% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-data-utils.R
\name{available_assays}
\alias{available_assays}
\title{Description of available assays in this packge}
\usage{
available_assays()
}
\value{
a tibble of assay information
}
\description{
Description of available assays in this packge
}
