% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-data-utils.R
\name{pb_split}
\alias{pb_split}
\title{Splits a SingleCellExperiment in a list of DGELists, split by the levels
one of the colData columns}
\usage{
pb_split(x, split.by = "cond")
}
\arguments{
\item{x}{A SingleCellExperiment or SummarizedExperiment}

\item{split.by}{a column name in `colData(x)` that has categorical covariate
(factor or character)}
}
\description{
This function is called from the dataset-wrangling.Rmd vignette, and
relies on the SingleCellExperiment package (which is in Suggests, not
Imports)
}
